Fix: models updated for API Compliance
Updated models to comply with API;
The following fields where not present in the expected response: allowCoupons, coupons, couponsUsed.

Also, AccountId and StoreId seem to not be sent by the API in some cases - Which implies they are optional or not used anymore;

Without these changes, the library is unusable.